# openapi.yaml
openapi: 3.1.0
info:
  title: OceanLaundry API
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  version: 0.3.0
  description: |
    Layanan laundry antar-jemput yang menghubungkan customer, kurir, dan staff
    outlet dalam satu alur: pemilihan paket, penjemputan, penimbangan &
    penetapan harga, pembayaran, pencucian, dan pengantaran Kembali.

    All monetary amounts are integers in the minor unit of Indonesian Rupiah (IDR).
    Clients MUST ignore unknown response fields.
servers:
  - url: https://oceanlaundry.api/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local development
security:
  - oauth2: [packages:read]
paths:
  /packages:
    get:
      summary: Get laundry packages
      operationId: getPackages
      responses:
        "200":
          description: List of available packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
        "401": 
          $ref: "#/components/responses/Unauthorized"
        "403": 
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "400":
          $ref: "#/components/responses/ValidationError"
    post:
      summary: Create a package
      operationId: createPackage
      security:
        - oauth2: [packages:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNewPackage"
      responses:
        "201":
          description: Package created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
        "401": 
          $ref: "#/components/responses/Unauthorized"
        "403": 
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
  /packages/{packageId}:
    get:
      summary: Get a package
      operationId: getPackage
      parameters:
        - $ref: "#/components/parameters/PackageId"
      responses:
        "200":
          description: Package details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
        "401": 
          $ref: "#/components/responses/Unauthorized"
        "403": 
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      summary: Update a package
      operationId: updatePackage
      security:
        - oauth2: [packages:write]
      parameters:
        - $ref: "#/components/parameters/PackageId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePackage"
      responses:
        "200":
          description: Updated package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
        "401": 
          $ref: "#/components/responses/Unauthorized"
        "403": 
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: Delete a package
      operationId: deletePackage
      security:
        - oauth2: [packages:write]
      parameters:
        - $ref: "#/components/parameters/PackageId"
      responses:
        "200":
          description: Deleted package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
        "401": 
          $ref: "#/components/responses/Unauthorized"
        "403": 
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /payments:
    post:
      summary: Make a payment
      operationId: createPayment
      security:
        - oauth2: [payments:write]
      description: Records a payment for an order. Transitions the order from
        awaiting_payment to washing. Returns 409 if the order is not in
        awaiting_payment status.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentRequest"
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "401": 
          $ref: "#/components/responses/Unauthorized"
        "403": 
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /payments/{paymentId}:
    get:
      summary: Get a payment
      operationId: getPayment
      security:
        - oauth2: [payments:read]
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "401": 
          $ref: "#/components/responses/Unauthorized"
        "403": 
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    oauth2:
      type: oauth2
      description: |
        Access tokens are issued by the OceanLaundry authorisation server.
        Public clients use Authorization Code + PKCE with no client secret.
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:8080/realms/ocean-laundry/protocol/openid-connect/auth
          tokenUrl: http://localhost:8080/realms/ocean-laundry/protocol/openid-connect/token
          refreshUrl: http://localhost:8080/realms/ocean-laundry/protocol/openid-connect/token
          scopes:
            packages:read: Melihat daftar package
            packages:write: Membuat, mengubah, menghapus package
            orders:read: Membaca order yang terlihat oleh pemanggil
            orders:write: Membuat dan membatalkan order milik sendiri
            orders:fulfil: Menimbang, memulai cuci, menandai siap, menandai selesai
            deliveries:write: Mengambil dan mengantarkan order
            payments:read: Membaca data pembayaran
            payments:write: Membuat pembayaran

    clientCredentials:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: http://localhost:8080/realms/ocean-laundry/protocol/openid-connect/token
          scopes:
            orders:read: Membaca order untuk mencari yang kedaluwarsa (automated job)
            orders:write: Membatalkan order yang kedaluwarsa (automated job)
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        UUID version 4 in canonical hyphenated form.
        Required on POST /v1/orders and POST /v1/payments.
        Keys are retained for 24 hours.
        If the same key is reused with an identical body, the server returns
        the previously stored response without processing the request again.
        If the same key is reused with a different body, the server returns
        409 Conflict with type
        https://oceanlaundry.api/problems/idempotency-key-reuse.
      schema:
        type: string
        format: uuid
    PackageId:
      name: packageId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    Package:
      type: object
      required: [id, name, description, price]
      properties:
        id:
          type: string
          example: pkg_001
        name:
          type: string
          example: Regular Wash
        description:
          type: string
          example: Cuci lipat 2 hari
        price:
          type: integer
          example: 8000
          minimum: 0
    CreateNewPackage:
      type: object
      required: [packageName, packageDesc, packagePrice]
      properties:
        packageName:
          type: string
          example: Regular Wash
        packageDesc:
          type: string
          example: Cuci lipat 2 hari
        packagePrice:
          type: integer
          example: 5000
          minimum: 0
    UpdatePackage:
      type: object
      properties:
        packageName:
          type: string
          example: Express Wash
        packageDesc:
          type: string
          example: Cuci lipat 1 hari
        packagePrice:
          type: integer
          example: 7000
          minimum: 0
...
    Payment:
      type: object
      required: [id, orderId, amount, status]
      properties:
        id:
          type: string
          example: pay_001
        orderId:
          type: string
          example: ord_001
        amount:
          type: integer
          description: Amount in minor unit of IDR.
          example: 28000
        status:
          type: string
          enum: [pending, paid, failed]
          example: paid
        createdAt:
          type: string
          format: date-time
          example: "2026-08-30T10:00:00Z"
    CreatePaymentRequest:
      type: object
      required: [orderId, amount]
      properties:
        orderId:
          type: string
          example: ord_001
        amount:
          type: integer
          description: Amount in minor unit of IDR.
          example: 20000
          minimum: 1
    Problem:
      type: object
      required: [type, title, status, detail, instance]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
  responses:
    Unauthorized:
      description: The token is absent, unverifiable, or expired.
      headers:
        WWW-Authenticate:
          schema: { type: string }
          example: 'Bearer error="invalid_token"'
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: The token is valid but does not carry the scope this operation requires.
      headers:
        WWW-Authenticate:
          schema: { type: string }
          example: 'Bearer error="insufficient_scope", scope="orders:fulfil"'
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: Request conflicts with the current state.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://oceanlaundry.api/problems/conflict
            title: Order state conflict
            status: 409
            detail: The requested transition is not allowed from the current order status.
            instance: /v1/orders/ord_001
    ValidationError:
      description: Invalid request.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://oceanlaundry.api/problems/validation-error
            title: Request validation failed
            status: 422
            detail: weightGrams must be greater than 0.
            instance: /v1/orders/ord_001/weigh
    NotFound:
      description: |
        The object does not exist, or it exists and is not accessible to this caller.
        Both conditions are answered identically so identifiers cannot be enumerated.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://oceanlaundry.api/problems/not-found
            title: Resource not found
            status: 404
            detail: The requested resource was not found.
            instance: /v1/orders/ord_001
    InternalServerError:
      description: Internal server error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://oceanlaundry.api/problems/internal-server-error
            title: Internal server error
            status: 500
            detail: An unexpected error occurred on the server.
            instance: /v1/orders/ord_001